app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pytz
itsdangerous
psycopg2-binary